requirements.txt
# Core dependencies for TAKexplorer
flask
waitress

# Database and data processing
# sqlite3 is built into Python 3.11+

# Google Cloud dependencies for App Engine
gunicorn  # Required for App Engine WSGI server
google-cloud-storage  # For database persistence

# Additional dependencies that TAKexplorer might need
# (add these based on the actual TAKexplorer requirements)
requests>=2.25.0
numpy>=1.19.0
